# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Nairobi Events Marketplace API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /users/me:
    get:
      operationId: getMe
      tags: [users]
      summary: Get current authenticated user
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "401":
          $ref: "#/components/responses/Unauthorized"
    patch:
      operationId: updateMe
      tags: [users]
      summary: Update current user profile
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateUserRequest"
      responses:
        "200":
          description: Updated user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "401":
          $ref: "#/components/responses/Unauthorized"

  # ─── Vendors ────────────────────────────────────────────────────────────────
...
  /quotes/requests:
    get:
      operationId: listMyQuoteRequests
      tags: [quotes]
      summary: List quote requests for authenticated vendor
      parameters:
        - name: status
          in: query
          schema:
            $ref: "#/components/schemas/QuoteStatus"
      responses:
        "200":
          description: Quote requests
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/QuoteRequest"
  /quotes/requests/{requestId}/submit:
    post:
      operationId: submitQuote
      tags: [quotes]
      summary: Vendor submits a quote for a request
      parameters:
        - name: requestId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SubmitQuoteRequest"
      responses:
        "201":
          description: Quote submitted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Quote"
  /quotes/{quoteId}/accept:
    post:
      operationId: acceptQuote
      tags: [quotes]
      summary: Planner accepts a quote (initiates booking)
      parameters:
        - $ref: "#/components/parameters/quoteId"
      responses:
        "200":
          description: Quote accepted, booking created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Booking"
  /quotes/{quoteId}/reject:
    post:
      operationId: rejectQuote
      tags: [quotes]
      summary: Planner rejects a quote
      parameters:
        - $ref: "#/components/parameters/quoteId"
      responses:
        "200":
          description: Quote rejected
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Quote"

  # ─── Bookings ────────────────────────────────────────────────────────────────
...
  /notifications:
    get:
      operationId: listNotifications
      tags: [notifications]
      summary: List notifications for current user
      parameters:
        - name: isRead
          in: query
          schema:
            type: boolean
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
      responses:
        "200":
          description: Notifications
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NotificationListResponse"
  /notifications/{notificationId}/read:
    patch:
      operationId: markNotificationRead
      tags: [notifications]
      summary: Mark a notification as read
      parameters:
        - name: notificationId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Notification marked read
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Notification"
...
components:
  parameters:
...
    quoteId:
      name: quoteId
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
  responses:
    Unauthorized:
      description: Authentication required
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
...
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
        message:
          type: string
      required: [error, message]

    # ─── Enums ────────────────────────────────────────────────────────────────
    UserRole:
      type: string
      enum: [planner, vendor, admin]
    EventType:
      type: string
      enum: [corporate, wedding, birthday, product_launch, conference, private_party, other]
    EventStatus:
      type: string
      enum: [draft, brief_submitted, quotes_requested, quotes_received, vendor_selected, booked, completed, cancelled]
    VendorCategory:
      type: string
      enum: [catering, mc, photography, videography, floristry, av_technical, tent_furniture, security, entertainment, decor, transportation, other]
    VendorStatus:
      type: string
      enum: [pending_review, approved, suspended, rejected]
    QuoteStatus:
      type: string
      enum: [requested, submitted, revision_requested, accepted, rejected, expired]
    BookingStatus:
      type: string
      enum: [pending, confirmed, in_escrow, completed, disputed, cancelled, refunded]

    # ─── User ─────────────────────────────────────────────────────────────────
    User:
      type: object
      properties:
        id:
          type: string
          format: uuid
        clerkId:
          type: string
        email:
          type: string
          format: email
        fullName:
          type: string
        phone:
          type: string
          nullable: true
        role:
          $ref: "#/components/schemas/UserRole"
        avatarUrl:
          type: string
          nullable: true
        isActive:
          type: boolean
        createdAt:
          type: string
          format: date-time
      required: [id, clerkId, email, fullName, role, isActive, createdAt]
    UpdateUserRequest:
      type: object
      properties:
        fullName:
          type: string
        phone:
          type: string

    # ─── Vendor ───────────────────────────────────────────────────────────────
    VendorProfile:
      type: object
      properties:
        id:
          type: string
          format: uuid
        userId:
          type: string
          format: uuid
        businessName:
          type: string
        category:
          $ref: "#/components/schemas/VendorCategory"
        description:
          type: string
        status:
          $ref: "#/components/schemas/VendorStatus"
        averageRating:
          type: string
          nullable: true
        totalReviews:
          type: integer
        totalBookings:
          type: integer
        isPremium:
          type: boolean
        city:
          type: string
        serviceAreas:
          type: array
          items:
            type: string
        websiteUrl:
          type: string
          nullable: true
        instagramHandle:
          type: string
          nullable: true
        portfolioUrls:
          type: array
          items:
            type: string
        createdAt:
          type: string
          format: date-time
      required: [id, userId, businessName, category, description, status, totalReviews, totalBookings, isPremium, city, createdAt]
...
    Event:
      type: object
      properties:
        id:
          type: string
          format: uuid
        plannerId:
          type: string
          format: uuid
        title:
          type: string
        eventType:
          $ref: "#/components/schemas/EventType"
        status:
          $ref: "#/components/schemas/EventStatus"
        eventDate:
          type: string
          format: date
        eventEndDate:
          type: string
          format: date
          nullable: true
        venue:
          type: string
          nullable: true
        city:
          type: string
        guestCount:
          type: integer
        budgetMin:
          type: string
          nullable: true
        budgetMax:
          type: string
          nullable: true
        currency:
          type: string
        servicesNeeded:
          type: array
          items:
            type: string
        specialRequirements:
          type: string
          nullable: true
        isEmergency:
          type: boolean
        quotesDeadline:
          type: string
          format: date-time
          nullable: true
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required: [id, plannerId, title, eventType, status, eventDate, city, guestCount, currency, servicesNeeded, isEmergency, createdAt, updatedAt]
...
    QuoteLineItem:
      type: object
      properties:
        description:
          type: string
        quantity:
          type: number
        unitPrice:
          type: number
        total:
          type: number
      required: [description, quantity, unitPrice, total]
    Quote:
      type: object
      properties:
        id:
          type: string
          format: uuid
        quoteRequestId:
          type: string
          format: uuid
        eventId:
          type: string
          format: uuid
        vendorId:
          type: string
          format: uuid
        category:
          $ref: "#/components/schemas/VendorCategory"
        status:
          $ref: "#/components/schemas/QuoteStatus"
        totalAmount:
          type: string
        currency:
          type: string
        depositPercent:
          type: integer
          nullable: true
        lineItems:
          type: array
          items:
            $ref: "#/components/schemas/QuoteLineItem"
        inclusions:
          type: array
          items:
            type: string
        exclusions:
          type: array
          items:
            type: string
        terms:
          type: string
          nullable: true
        notes:
          type: string
          nullable: true
        version:
          type: integer
        submittedAt:
          type: string
          format: date-time
        vendor:
          $ref: "#/components/schemas/VendorProfile"
      required: [id, quoteRequestId, eventId, vendorId, category, status, totalAmount, currency, lineItems, version, submittedAt]
    QuoteRequest:
      type: object
      properties:
        id:
          type: string
          format: uuid
        eventId:
          type: string
          format: uuid
        vendorId:
          type: string
          format: uuid
        category:
          $ref: "#/components/schemas/VendorCategory"
        status:
          $ref: "#/components/schemas/QuoteStatus"
        requestedAt:
          type: string
          format: date-time
        expiresAt:
          type: string
          format: date-time
        event:
          $ref: "#/components/schemas/Event"
      required: [id, eventId, vendorId, category, status, requestedAt, expiresAt]
    SubmitQuoteRequest:
      type: object
      properties:
        totalAmount:
          type: string
        depositPercent:
          type: integer
        lineItems:
          type: array
          items:
            $ref: "#/components/schemas/QuoteLineItem"
        inclusions:
          type: array
          items:
            type: string
        exclusions:
          type: array
          items:
            type: string
        terms:
          type: string
        notes:
          type: string
        validUntil:
          type: string
          format: date-time
      required: [totalAmount, lineItems]

    # ─── Booking ──────────────────────────────────────────────────────────────
    Booking:
      type: object
      properties:
        id:
          type: string
          format: uuid
        eventId:
          type: string
          format: uuid
        quoteId:
          type: string
          format: uuid
        vendorId:
          type: string
          format: uuid
        plannerId:
          type: string
          format: uuid
        status:
          $ref: "#/components/schemas/BookingStatus"
        totalAmount:
          type: string
        platformFeeAmount:
          type: string
        vendorPayoutAmount:
          type: string
        currency:
          type: string
        stripePaymentIntentId:
          type: string
          nullable: true
        contractUrl:
          type: string
          nullable: true
        confirmedAt:
          type: string
          format: date-time
          nullable: true
        completedAt:
          type: string
          format: date-time
          nullable: true
        createdAt:
          type: string
          format: date-time
      required: [id, eventId, quoteId, vendorId, plannerId, status, totalAmount, platformFeeAmount, vendorPayoutAmount, currency, createdAt]
...
    Notification:
      type: object
      properties:
        id:
          type: string
          format: uuid
        userId:
          type: string
          format: uuid
        type:
          type: string
        title:
          type: string
        body:
          type: string
        isRead:
          type: boolean
        readAt:
          type: string
          format: date-time
          nullable: true
        createdAt:
          type: string
          format: date-time
      required: [id, userId, type, title, body, isRead, createdAt]
    NotificationListResponse:
      type: object
      properties:
        notifications:
          type: array
          items:
            $ref: "#/components/schemas/Notification"
        unreadCount:
          type: integer
      required: [notifications, unreadCount]

    # ─── Admin ────────────────────────────────────────────────────────────────
...