# lib/api-zod/src/generated/types/quoteRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Nairobi Events Marketplace API
 * OpenAPI spec version: 0.1.0
 */
import type { Event } from "./event";
import type { QuoteStatus } from "./quoteStatus";
import type { VendorCategory } from "./vendorCategory";

export interface QuoteRequest {
  id: string;
  eventId: string;
  vendorId: string;
  category: VendorCategory;
  status: QuoteStatus;
  requestedAt: Date;
  expiresAt: Date;
  event?: Event;
}