# lib/api-zod/src/generated/types/submitQuoteRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Nairobi Events Marketplace API
 * OpenAPI spec version: 0.1.0
 */
import type { QuoteLineItem } from "./quoteLineItem";

export interface SubmitQuoteRequest {
  totalAmount: string;
  depositPercent?: number;
  lineItems: QuoteLineItem[];
  inclusions?: string[];
  exclusions?: string[];
  terms?: string;
  notes?: string;
  validUntil?: Date;
}