# lib/api-zod/src/generated/types/userRole.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Nairobi Events Marketplace API
 * OpenAPI spec version: 0.1.0
 */

export type UserRole = (typeof UserRole)[keyof typeof UserRole];

export const UserRole = {
  planner: "planner",
  vendor: "vendor",
  admin: "admin",
} as const;